// src/generated/models/Geek_project_tasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Geek_project_tasks {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  geek_category?: string;
  geek_categoryname?: string;
  geek_description?: string;
  geek_duration?: number;
  geek_end?: Date;
  geek_id?: string;
  geek_lookup_assignee?: string;
  geek_lookup_assigneename?: string;
  geek_lookup_assigneeyominame?: string;
  geek_name?: string;
  geek_priority?: string;
  geek_priorityname?: string;
  geek_progress?: number;
  geek_project_taskid: string;
  geek_projectid?: string;
  geek_projectidname?: string;
  geek_start?: Date;
  geek_status?: string;
  geek_statusname?: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}